Changes after code review part 2
