+++ src/server/routes/controllers/logInUser.js
@@ -1,10 +1,10 @@
-const logInUtilities = require('./utilities/logInUtilities')
+const initializeLoginHandler = require('./utilities/initializeLoginHandler')
 const strategyUtils = require('../../passport/strategyUtils')
 
 function logIn(passport = require('passport')) {
 	return (req, res, next) => {
-		const proceedLogIn = logInUtilities(req, res, next).proceedLogIn
-		const logInByPassport = passport.authenticate(strategyUtils.STRATEGY_NAME, proceedLogIn) 
+		const loginHandler = initializeLoginHandler(req, res, next)
+		const logInByPassport = passport.authenticate(strategyUtils.STRATEGY_NAME, loginHandler) 
 		logInByPassport(req, res, next)
 	}
 }
